refactor: restructure directory listing download module
