+++ src/settings/file.rs
@@ -16,7 +16,7 @@ use std::{collections::BTreeSet, path::PathBuf};
 use crate::directory_listing::DirListFmt;
 
 #[cfg(feature = "directory-listing-download")]
-use crate::directory_listing_download::DirDownloadFmt;
+use crate::directory_listing::download::DirDownloadFmt;
 
 use crate::{Context, Result, helpers};
